+++ packages/react/FormField/src/components/FormFieldSelect/FormFieldSelect.tsx
@@ -26,6 +26,4 @@ const FormFieldSelect: React.FC<FormFieldSelectProps> = ({
   </FormField>
 );
 
-FormFieldSelect.displayName = "FormField.Select";
-
 export { FormFieldSelect };