+++ packages/react/FormField/src/components/FormFieldTextarea/FormFieldTextarea.tsx
@@ -26,6 +26,4 @@ const FormFieldTextarea: React.FC<FormFieldTextareaProps> = ({
   </FormField>
 );
 
-FormFieldTextarea.displayName = "FormField.Textarea";
-
 export { FormFieldTextarea };