+++ packages/react/Menu/src/components/MenuBody/MenuBody.tsx
@@ -26,6 +26,4 @@ const MenuBody: React.FC<MenuBodyProps> = ({
   </Box>
 );
 
-MenuBody.displayName = "Menu.Body";
-
 export { MenuBody };