+++ packages/react/Menu/src/components/MenuFooter/MenuFooter.tsx
@@ -30,6 +30,4 @@ const MenuFooter: React.FC<MenuFooterProps> = ({
   );
 };
 
-MenuFooter.displayName = "Menu.Footer";
-
 export { MenuFooter };