+++ packages/react/Menu/src/components/MenuHeader/menuHeader.types.ts
@@ -1,5 +1,6 @@
 import { HTMLAttributes, ReactNode } from "react";
 
 export interface MenuHeaderProps extends HTMLAttributes<HTMLElement> {
+  /** Content of the menu header */
   children: ReactNode;
 }