+++ packages/react/ProductUpdates/src/productUpdates.types.ts
@@ -9,8 +9,12 @@ type Extend = Omit<
   HTMLAttributes<HTMLElement>;
 
 export interface ProductUpdatesProps extends Extend {
+  /** Title of the Popover */
   title: string;
+  /** Content text of the Popover */
   text: string;
+  /** Optional content for the body of the Popover */
   bodyContent?: ReactNode;
+  /** Optional bottom link to dismiss the Popover. If unset, an IconButton with an X will appear on the superior right corner. */
   dismissLink?: ReactNode;
 }