+++ packages/react/Page/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Page component allows the user to build application pages with all the necessary attributes for constructing a variety of layouts.
 
+## 2023-03-13 `1.1.1`
+
+### 📚 Documentation
+
+- Add props documentation. ([#26](https://github.com/TiendaNube/nimbus-patterns/pull/26) by [@juanchigallego](https://github.com/juanchigallego))
+
 ## 2023-02-17 `1.1.0`
 
 ### 🎉 New features